pyproject.toml
[tool.poetry]
name = "hannah-tvm"
version = "0.1.0"
description = ""
authors = ["Christoph Gerum <christoph.gerum@uni-tuebingen.de>"]
license = "Apache-2.0"
#build = "build.py"

[tool.poetry.dependencies]
python = ">3.7.1, <3.11"
hydra-core = "^1.1.0"
onnx = "^1.8.1"
importlib-metadata = "^4.0.1"
Twisted = "^21.2.0"
prettytable = "^2.1.0"
SQLAlchemy = "^1.4.12"
configparser = "^5.0.2"
pyvcd = "^0.2.4"
pyelftools = "^0.27"
openpyxl = "^3.0.7"
XlsxWriter = "^1.4.0"
PyYAML = "^5.4.1"
numpy = "^1.20.2"
tflite = "^2.4.0"
colorama = "^0.4.4"
decorator = "^5.1.0"
tqdm = "^4.62.3"
torch = "^1.13.1"
networkx = ">=2.6.3"
matplotlib = "^3.4.3"
tensorflow = ">=2.7.0"
graphviz = ">=0.19.1"

pyserial = {version=">=3.5", optional=true}
usb = {version=">=0.0.83.dev0", optional=true}
west =  {version=">=0.13.1", optional=true}
onnxoptimizer = ">=0.2.6"
xgboost = "1.5.2"
dash = {version="^2.3.1", optional=true}
pandas = "<1.4.0"
tabulate = ">=0.8.9"
PTable = {version=">=0.9.2", optional=true}
board-automate = {version=">=0.2.0", optional=true, extras=["postgres"]}
fsspec = {version = "^2022.8.2", extras=["git", "http"]}
appdirs = "^1.4.4"
typing-extensions = "^4.4.0"


[tool.poetry.scripts]
hannah-tvm-compile = 'hannah_tvm.compile:main'
hannah-tvm-tune = 'hannah_tvm.tune:main'
hannah-tvm-memory = 'hannah_tvm.analysis:memory_main'
hannah-tvm-dashboard = 'hannah_tvm.dashboard.app:main'

[tool.poetry.extras]
automate = ["board-automate"]
micro = ["pyserial", "usb", "PTable", "west"]
dash = ["dash"]

[tool.poetry.dev-dependencies]
black = ">=22.3.0"
pre-commit = ">=2.12.1"
pytest = ">=6.2.5"
isort = ">=5.10.1"
coverage = ">=6.3.2"
pyment = "^0.3.3"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu113"
default = false
secondary = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
